+++ src/components/Routes/Router.js
@@ -1,7 +1,7 @@
 import { Suspense, } from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { SuspenseLoader } from "./Suspense";
-import { AdminLayout, Dashboard, FundingPage, KitPage, Login, MonthlyRenewalPage, RenewalPage, Requests, SignUp, UserDetails, Users, Wallet, WalletHistory, WithdrawalsList } from "./LazyLoader";
+import { AdminLayout, Dashboard, FundingPage, KitPage, Login, MonthlyRenewalPage, RenewalPage, Requests, SignUp, UserDetails, Users, Wallet, WalletHistory } from "./LazyLoader";
 import InvoicesPage from "../../pages/Invoice";
 
 
@@ -37,7 +37,6 @@ const AppRouter = () => {
                 <Route path="/renewals" element={<RenewalPage />} />
                 <Route path="/wallet" element={<Wallet />} />
                 <Route path="/wallet-history" element={<WalletHistory />} />
-                <Route path="/wallet-withdrawal" element={<WithdrawalsList />} />
                 <Route path="/kits" element={<KitPage />} />
                 <Route path="/users" element={<Users />} />
                 <Route path="/invoices" element={<InvoicesPage />} />
